"9:55am. My wakeup time was better than yesterday. Still, I think I'll follow the same schedule. I want to pack winter preps away and I can do it all today, so my plan is to have brekafast now, then get to work on that. After that I'll get to work on getting back to par.
10:35am. Done with breakfast. The chores are next. I'll try to get it out of way entirely today.

4pm. I've come back roughly 5m ago and have just finished doing over 5h of manual labour. Before I can resume programming, let me take a break.

On the plus side I am done with winter prep. No more of this. From here on out and for the rest of the winter I will be free.

4:35pm. Done with lunch. Usually I have it at a later time, but I was pretty hungry. Now, obviously I do not have time to accomplish much in the rest of the day, but how about I focus on doing something? That is the way to go about this.

Let me put the `package_update` into action. I should have time for that.

4:40pm. Hmmm, no. I can't just plug it in. I need to rethink everything.

Today I did have 5.5 hour in which to think about what I am going to do today, but I haven't spared even a thought to it. Because programming is a burden and I do not feel like dealing with two different kinds of work. In the future when Spiral is done and I have neurochips with which to play with, my enthusiasm for it should spring back to life, but for now it can only be described as work and I am not hard working enough to deal with two things.

Nonetheless, for the next hour or so, until 6pm, I will direct my thoughts to the task at hand.

Let me see if I can do some quality thinking. I need to figure out what the next step should be.

```fs
type SupervisorReq =
    | ProjectFileOpen of {|uri : string; spiprojText : string|}
    | ProjectFileChange of {|uri : string; spiprojText : string|}
    | ProjectFileDelete of {|uri : string|}
    | ProjectFileLinks of {|uri : string|} * RString list IVar
    | ProjectCodeActions of {|uri : string|} * RAction list IVar
    | ProjectCodeActionExecute of {|uri : string; action : ProjectCodeAction|}
    | FileOpen of {|uri : string; spiText : string|}
    | FileChanged of {|uri : string; spiEdit : SpiEdit|}
    | FileTokenRange of {|uri : string; range : VSCRange|} * VSCTokenArray IVar
    | HoverAt of {|uri : string; pos : VSCPos|} * string option IVar

let supervisor tokenizer_errors parser_errors type_errors package_errors fatal_errors req =
    let loop = req >>= function
        | ProjectFileOpen x -> failwith "TODO"
        | ProjectFileChange x -> failwith "TODO"
        | ProjectFileDelete x -> failwith "TODO"
        | ProjectFileLinks(x,res) -> failwith "TODO"
        | ProjectCodeActions(x,res) -> failwith "TODO"
        | ProjectCodeActionExecute x -> failwith "TODO"
        | FileOpen x -> failwith "TODO"
        | FileChanged x -> failwith "TODO"
        | FileTokenRange(x, res) -> failwith "TODO"
        | HoverAt(x,res) -> failwith "TODO"
    Job.foreverServer loop
```

Yeah, I have no choice but to start work on this. Yesterday I wrote that it is goals that makes the earth and heavens move.

Right now my problem is that I do not know what to do with the old code in `OldServer.fs`.

5pm. So it is best that I clear my mind and start by rebuilding. I'll go through all the reasoning from scratch, taking advantage of what I've done previously. This is the way to do redesigns. Just grab whatever you need, but don't be bound by it. Not being bound does imply one needs to be willing to wreck everything.

Let me do it. Let me do the `ProjectFileOpen` and the rest.

For the next hour, I should just focus on that.

5:05pm. Focus me.

First thing's first - let me make the default supervisor state.

```fs
    let mutable s : SupervisorState = {
        modules = Map.empty
        packages = {
            validated_schemas = Map.empty
            package_links = mirrored_graph_empty
            packages = Map.empty
            }
        }
```

Yeah, this is the sensible first step.

```fs
| ProjectFileOpen x -> failwith "TODO"
```

Now can I do the open case? Let me just focus on this.

5:30pm.

```fs
    let loop s = req >>= function
        | ProjectFileOpen x ->
            let dir = dir x.uri
            let s = package_update errors s dir (Some x.spiprojText)
            let order,packages = package_validate s.packages dir
            package_errors order packages |> Array.iter (fun er ->
                Hopac.start (Src.value errors.package er)
                )
            Job.result {s with packages=packages}
```

My foucs is not exactly that high, but I managed to get this out.

5:35pm. Since I am losing my thread, let me try to get the next two cases done as well.

```fs
    let loop s = req >>= function
        | ProjectFileChange x | ProjectFileOpen x ->
```

Actually, let me just merge these two case. Who gives a crap if a bit extra work is done every time the user opens a project file. I do not need to optimize this.

```fs
| ProjectFileDelete x -> failwith "TODO"
```

Let me do this one.

5:45pm.

```fs
let package_validate_then_send_errors errors s dir =
    let order,packages = package_validate s.packages dir
    package_errors order packages |> Array.iter (fun er ->
        Hopac.start (Src.value errors.package er)
        )
    {s with packages=packages}

let supervisor (errors : SupervisorErrorSources) req =
    let loop s = req >>- function
        | ProjectFileChange x | ProjectFileOpen x ->
            let dir = dir x.uri
            let s = package_update errors s dir (Some x.spiprojText)
            package_validate_then_send_errors errors s dir
        | ProjectFileDelete x ->
            let dir = dir x.uri
            let s = {s with packages={s.packages with validated_schemas=Map.add dir (Error "The package file does not exist.") s.packages.validated_schemas}}
            package_validate_then_send_errors errors s dir
```

How about this?

5:50pm. Yeah, I think this is quite fine. Previously on project file deletion I tried reloading it just to make sure, but now I am doing what I need directly.

5:55pm. Am doing some refreshing to make sure that the validation would get done on removal. It will. No problem.

```fs
        | ProjectFileLinks(x,res) -> failwith "TODO"
        | ProjectCodeActions(x,res) -> failwith "TODO"
        | ProjectCodeActionExecute x -> failwith "TODO"
```

The next step are these. I should be able to fill these out quickly. Let me do it and I'll call it a day.

6pm.

```fs
        | ProjectFileLinks(x,res) ->
            match Map.tryFind (dir x.uri) s.packages.package_schemas with
            | Some (Ok x) -> Hopac.start (IVar.fill res (List.append x.schema.links x.package_links))
            | _ -> Hopac.start (IVar.fill res [])
            s
        | ProjectCodeActions(x,res) ->
            match Map.tryFind (dir x.uri) s.packages.package_schemas with
            | Some (Ok x) -> Hopac.start (IVar.fill res x.schema.actions)
            | _ -> Hopac.start (IVar.fill res [])
            s
        | ProjectCodeActionExecute x ->
            match code_action_execute x.action with
            | Some er -> Hopac.start (Src.value errors.fatal er)
            | None -> ()
            s
```

This is good enough for the day. Let me shut down here. I want to have some fun. Tomorrow I will focus properly on dealing with modules. I'll deal with having proper package updates as well.

Those package updates are a weight on my mind. I really want to get them out of the way as soon as possible."