"2:15pm. Ok, Mahoako chapter, an ep of Elaina, chores and then I will resume.
3:20pm. I did not even got to watch that ep, but nevermind. Let me resume.

2.5 hours of slacking is enough for me.

Let me focus on finishing the multi-file stream for the prepass.

3:35pm.

```fs
let multi_file package_id top_env =
    let rec create files' =
        {new MultiFileStream with
            member _.Run files =
                let files = diff_order_changed files' files
                let x, l = multi_file_run top_env_empty prepass id Prepass.union Prepass.in_module package_id top_env files
                snd x, create l
            }
    create []
```

Yeah, this is it. It compiles now. I renamed `Stream` to `EditorStream` because F#'s editor support was acting wonky. It keep confusing it for the core stream for some reason.

3:40pm. Now I am just getting lost in thought again.

How about I commit here? What I have here is definitely the multi file stream for the prepass.

I need to do some thinking how to deal with the packages. For one, unlike for the typeching phase I won't need the diff stream..."