"2:20pm. Done with resting. Let me start.
It is time to get things done.

Let me move `package_multi_file` to the typer stream module.

```fs
| None -> (package_multi_file (fun _ x -> x) multi_file package_env_default union in_module top_to_package package_to_top).Run(id,links,files)
```

And now this comes out like so. Piece of cake.

2:30pm. I am just doing a little review.

...Yeah, things are very nice right now.

I have the package stream for the prepass done.

No doubt about that.

Let me dwell on this just for a bit and then I will get started on the review. I meant to do it yesterday, but I wanted to get these tasks done before.

```fs
                    let f package_env_in env_out = env_out >>=* fun env_out -> package_env_in >>- fun package_env_in -> top_to_package id env_out package_env_in
                    let env_out = if top_env_out = top_env_out' then env_out' else f package_env_in top_env_out
```

I can simplify this a bit.

```fs
                    let f env_out = env_out >>=* fun env_out -> package_env_in >>- fun package_env_in -> top_to_package id env_out package_env_in
                    let env_out = if top_env_out = top_env_out' then env_out' else f top_env_out
```

2:35pm. Ok, let me leave that aside.

Doing these package streams and factoring out the code between them and the typer streams was wonderful work. If anybody is wondering whether I know anything about concurrency, I can just point them at this in the future.

All the struggles were so I could make these. I trully exercised the entirety of my skill in their making.

And now I am pretty decent as far as concurrency is concerned.

The way the package loader and the attention servers work is enlightened as well.

2:45pm. Let me commit here and I will turn off the router as I've already started browsing social media. It is an impulse that triggers whenever I have nothing to do.

The next thing on my list is to look inwardly and do that review. Let me do that for an hour or two, and maybe I'll call it a day depending on how long it takes me to do it.

Programming-wise, the only thing I have in mind is to do the function that takes in the `SupervisorState` and calls the package stream. That won't be hard to do, but it will require some time and focus."