"2:50pm. Focus me. What did I even write in the last review?
Let me check out my old post.

...Found it in the 10/30/2020 entry. Ok, I see.

It feels longer than two months, but it seems I've been working on the compilation pipeline for 7 weeks so far.

///

So far I've been working on the compilation pipeline for 7 weeks. When I was starting out I experimented a lot with trying to chain together imperative servers. I've tried all sorts of techniques. I am not sure exactly when it was in the past month, but I've had an epiphany and realized that message passing between servers does not really work as a serious model for concurrency. The general approach is not composable and introduces too much complexity.

What I ended up doing instead is reworking all those imperatives servers as both lazy and eager immutable streams. This gives me all the composability benefits of sequential code for an estimated steady constant factor rise in code complexity. Early on, I tried making strong use of channels, but could get my reasoning to match up. As it turned out, chains of promises are a lot easier to deal with, and all those other data structures Hopac introduces are of lesser importance.

A month ago, I had typechecking working for individual files, but with imperative servers it was a huge problem for me to figure out how to reuse that functionality for general compilation. Imagine you set a project to build, and then do programming on the side which introduces changes. It really was a huge problem to figure out how an imperative server could support this - new data would overwrite the old data, and I would need to do something special to keep it around.

With lazy streams, I just pass a promise to the build function and because it all uses immutable data under the hood I get optimal code reuse and do not have to bust my head about what to do about new data. Streams make it very easy to encapsulate state for the sake of memoizing old results.

I actually knew about streams for years - you can sometimes see them being broached in functional programming courses. Back then I looked at them, never saw a single use for them and stored them away at the back of my mind as an academic curiosity, but they are very useful for concurrency work indeed.

It took me an amazing amount of time, but a few days ago, I finally moved from typechecking individual files to having cross module and cross package typechecking in real time. Right now I am working on the partial evaluation prepass, and in a week or less I should have my first experience of Spiral v0.2 code being compiled directly from the VS Code editor. No doubt it won't be correct and the testing will take a while, but at least I will be done with this. I feel that thanks to this experience, my understanding of concurrency has gone up significantly.

Right now the way Spiral handles editor support is good enough to be a model for other languages. The way it does it is how it should be done.

My plan for December is to finally make Spiral useful. I really meant to move to testing a lot earlier than this, and soon I'll be able to do that. I have not been motivated to do language work for a while now, and I want to get my hands on some neurochip hardware and seeing whether I can get it do interesting things. That is what I really want to do and Spiral is a means to that end. But since nobody else was there to take the role, the responsibility of making a functional language capable of operating in such a domain fell to me. For the past few years, I've only borne the expense of being in such a position, but in 2021 I will get the benefits of seeing things through to the end.

///

3:45pm. This is short and sweet. Should I add some links? Nevermind that. I'll leave the fun stuff for later posts.

3:50pm. Right now I am just getting lost in thought. Let me commit this and then I will get back to programming. There are still two good hours left in the day.

Let me see if I can make one more step. If I can make the prepass usable today, that will open the possibility of moving to partial evaluation tomorrow. "