"10:55am. Let me start. That rigged duel of Valerie's messed up my schedule. I just could not let it go. I should not have given her the letter and got rid of the paladin on my own.
Let me move to testing.

11am. Hmmm, compilation is failing, but there aren't any errors.

11:05am. Gahhhh, how annoying. F# just gets worse as time goes on it seems.

...Restarting the IDE did the trick. I suppose I could have tried clearing the build as well.

Let me finally run this. Before I try actual compilation, let me make sure that regular typechecking still works after the flurry of changes that I did.

...Ah, no wait. It seems I am getting an exception right at start up. What the hell.

```fs
    Job.iterateServer {
        modules = Map.empty
        diff_stream = package_diff
        infer_results = Map.empty
        package_ids = PersistentHashMap.empty
        packages = {
            validated_schemas = Map.empty
            package_links = mirrored_graph_empty
            package_schemas = Map.empty
            }
        prepass_stream = Prepass.package
        } loop
```

It is saying it is here, and that the given key 0 is not in the dictionary.

```
KeyNotFoundException: The given key '0' was not present in the dictionary.
```

```
This exception was originally thrown at this call stack:
    [External Code]
    Spiral.PartEval.Prepass.process_ty@541-27T<b>.Invoke(Spiral.PartEval.Prepass.LowerEnv, b) in Prepass.fs
    Spiral.PartEval.Prepass.process_ty@541-37.Invoke(Spiral.PartEval.Prepass.LowerEnv, Spiral.PartEval.Prepass.T) in Prepass.fs
    Spiral.PartEval.Prepass.process_ty(Spiral.PartEval.Prepass.T) in Prepass.fs
    Spiral.PartEval.Prepass.prepass@778.Invoke(Spiral.PartEval.Prepass.T) in Prepass.fs
```

What is this?

```fs
let process_ty (x : T) =
    let scope = propagate (Choice2Of2 x)
    resolve scope (Choice2Of2 x)
    lower scope (Choice2Of2(x,id))
```

So it is here.

```fs
let top_env_default =
    let rec f (m : PersistentHashMap<string,int>) = function
        | TyVar x -> TV m.[x.name]
        | TyPrim x -> TPrim x
        | TyArray x -> TArray (f m x)
        | TyLayout(a,b) -> TLayout(f m a,b)
        | TyInl(a,b) -> TArrow(m.Count,f (m.Add(a.name,m.Count)) b)
        | _ -> failwith "Compiler error: The base type in Infer is not supported in the prepass yet."

    List.fold (fun (top_env : PrepassTopEnv) (k, x) ->
        {top_env with ty = Map.add k ((prepass -1 0 "<base_types>" top_env).base_type (f PersistentHashMap.empty x)) top_env.ty}
        ) top_env_empty Infer.base_types
```

Ah, no wait. I should be looking into it here.

11:15pm. F#'s debugger is being an ass, and not letting me step into properly.

```fs
        | TyInl(a,b) -> TArrow(m.Count,f (m.Add(a.name,m.Count)) b)
        | _ -> failwith "Compiler error: The base type in Infer is not supported in the prepass yet."

    List.fold (fun (top_env : PrepassTopEnv) (k, x) ->
        printfn "Trying %s" k
        let x = (prepass -1 0 "<base_types>" top_env).base_type (f PersistentHashMap.empty x)
        printfn "Done..."
```

This happens why I try doing TArray.

```
TArrow (0, TArray (TV 0))
```

It is trying to eval this expression. What happens when it tries doing that?

```fs
let process_ty (x : T) =
    let scope = propagate (Choice2Of2 x)
    resolve scope (Choice2Of2 x)
    lower scope (Choice2Of2(x,id))
```

Let me get rid of the inline annotation on these 3. Debugger + inlining does not work well with F#.

11:30am.

```fs
        | TArrow(a,b) ->
            let scope, env = scope env a
            TArrow'(scope,adj' env a,ty env b)
```

I am starting to get a handle on this. Basically, it goes here and then...

```fs
    let scope (env : LowerEnv) x =
        let v = scope.[x]
```

It tries getting something that is not in the env and throws an exception.

11:45am.

```fs
        | TArrow(a,b) ->
            let scope, env = scope env a
            TArrow'(scope,adj' env a,ty env b)
```

I've just been looking at this trying to refresh my memory of what this entire function was supposed to be doing in the first place.

I think the possible issue is that I doing `scope env a` instead of `scope env x`.

```fs
        | TArrow(a,b) ->
            let scope, env = scope env x
            TArrow'(scope,adj' env a,ty env b)
```

Yeah, let me try this. What I had previously was definitely a mistake.

I haven't ever tried running the prepass before now, so it is no wonder strange things are happening.

Yeah, now the exception is gone.

11:55am. I am trying out whether the things I have previously still work and the answer is no. There is a lag of one keypress between recognizing an error. In addition to that the hovers outright fail to work for some files. Quite strange.

Let me stop here for breakfast.

12pm. These weird things always happen after a long period of change without testing. Thankfully, now that I have the whole language, I can do integration testing properly and won't have to write code for weeks without ever running it. I know this is not a good habit, but it makes no sense for me to test like half a pass.

Even for something like testing the prepass, I first need to parse and typecheck everything first. It is hard to do proper testing without meeting the prerequisites first.

Starting here and for the next few days or weeks, I will be hunting down bug after bug until everything is working as it should. This will take some effort, but it should not take too long.

After breakfast, I think I will go back to a previous commit and see whether I had the same error previously. Alternatively, I can just look for the problem proactively.

12:05pm. No matter which choice I pick, the effort is what matters."