"9:55am. As a reward for myself, let me take a small vacation today. Maybe I'll skip the morning session. I went to bed really late, but I still got up at this time. Let me chill a bit, and then I will talk about what is on my mind.
10:45am. I am almost done having my fun.

10:55am. I'll save the Toji No Miko game subs for later. Let me gather my thoughts.

Ever since I closed for the day yesterday, my energy level was very high to the point of interfering with my sleep, so I am letting some of it loose by slacking.

11:05am. I am thinking. Let me just write a thought down. This wasn't what I wanted to rant about, but since it is at the forefront of my mind, let me do it.

```fs
    use client = new RequestSocket()
    client.Connect(uri_client)
```

I am should make this a pub socket. Looking at the supervisor now, I think I am mostly set, but a request socket can only interface with a single server. I can't really use it to send errors to more than one destination.

11:25am. Had to step away for a bit, let me resume my train of thought.

I remember thinking in the past that VS Code's providers should really have just taken in observables, but I was wrong. Having providers send requests themselves means that a single language server can serve multiple editor instances.

If VS Code had the design I'd have suggested, I would be in real trouble now.

As things stack, I am in a really good position at the moment. All I need to do is change the request socket to a pub one, and the response on to a sub one on the VS Code side and I am good.

This is remarkable. I though about this months ago, and making Spiral support multiple editor instances within a single server seemed a lot more difficult back then. I forgot what the reason for that was. It does not matter.

11:30am. After changing the sockets, the only real change I need to make in order to make the plugin publication ready is to have it start the server on its own. I'll try it with a hardcoded path at first.

Hmmm...since the user won't have the command line open on the side, it would not be bad if Hopac's uncaught exceptions could be rerouted as fatal errors in the editor.

11:35am. Well, after that, there is also the matter of publishing the language. I am not familiar with how this should be done on the VS Code marketplace.

I think since I am going to be doing this, it is finally time that I clean up the repo of all the previous stuff and merge it with master. Actually, I'll merge and then clean up. Yesterday when I tried doing CFR I was surprised to find my memory scrambled. But even though I've forgotten where I put some things, the stuff in `Games`, `Learning` and `Cuda` folder is still valuable. I want to leave it around in some repo.

In fact, the poker game that is in Games might be a worthy project of porting to v0.2.

It is just, the whole solution has too much stuff in it right now, and I want to start with a clear slate in master. I'll get rid of all the pre-v0.2 stuff and clean up some of the v0.2 stuff as well.

11:50am. I am thinking. All the goals I have in mind are easy to describe, but some like publishing the plugin and packaging it with the language will require research time. I won't rush this. It probably won't take this long, but I'll give myself a few days to get this out of the way.

11:55am. I never thought I'd be doing it this soon, so it is no wonder I am excited. I was expecting that I would be doing this months from now.

But it is not like the language is unusable in its current form. I'll say it is prerelease alpha, and test it while it is out along with filling out the new documentation.

12pm. Right...it is good to spend some time just planning things out. Desires without steps to achieve them are just empty dreams. But dreams with steps to attain them will eventually become reality. You first desire the dream, then you forget the dream and desire the steps themselves.

12:05pm. The other point I want to dwell on is something I've mentioned yesterday.

One strong desire I have compared to before is not to do things from the command line.

This is actually my main anxiety right now. I do not want to drop Avalonia, Hopac and Rx just because I will be moving to Spiral.

I want to cement this desire. I will absolutely not give up these advantages that were so hard to get. I will not go back to being a command line loser again. I am not sure what I will do, but I will make it my priority to be able to do concurrency and GUIs from Spiral, whether that means having bindings to these libraries or figuring out how to do a hybrid approach.

12:15pm. Forget the CFR. Forget ML. Those things I've already done 100s of times before.

What I should be doing are GUIs. Those should be my next project in Spiral. ML is easy.

It used to be the hard part - but I have Spiral. It is not going to be an obstacle. For the next few months, what I should be my main aim is getting comfortable with making ML interactive.

That poker game in `Games` - why don't I make it interactive? Why not do a poker game in the way that it should be done?

Interactivity will be important every step of the way - for the sake of training agents, I will need my control center. I will need interactive games in order to play against the agents myself. And once I have good agents, I am not going to be directing their attacks against the gamling dens from the command line.

12:20pm. In fact...let me elaborate those ideas.

Back in late 2016, I really meant Spiral to be a code generator for an F# ML library.

Now that Spiral is no longer a run-and-done compiler, but a memoizing server, it might be worth considering that vision again.

I haven't really figured out yet how to make use of the F# compiler from F# itself. But there are possibilities.

12:25pm. Months ago, I've seen in preview that Fabulous can do UIs with hot-reloading. I need to explore possibilities in this direction.

There is nothing wrong with scripting languages - Spiral is fast, but performance is a means to an end. A whole program optimizing compiler like it is ideal for doing the backend. It would be great for compiling the core logic of my agents.

But I should not be obligated to do everything in it. What I am obligated is to use it to maximum possible effect.

12:30pm. Interafaces - I am not done with them. My work in the Lithe repo was just the first taste. And the concurrency work I did this year was just the first taste as well. I finally know how to do this properly. I need to go further. I need to keep exercising those skills. I've just made them a part of the new foundation.

12:35pm. Oh yeah. I forgot something - if there are multiple VS Code instances, which one own the server? When should the server close?

I actually don't know how to deal with that at all. I thought about this before, but drew a blank. I know how to check if an instance exists, but not more than that.

Let me stop here for the morning."