"10:30am. I am finally getting up earlier even if by only a little. Yesterday exhausted me. Writing that last review opened my eyes.
Whether my plan succeeds or not really depends on the competence level of the teams doing the software stacks at those hardware companies. Assuming there is enough choice for me to pick my targets, I need to focus primarily on those with functional programming experience. Alternatively, those open to learning new languages. Otherwise things will get really hard.

10:45am. These sorts of worries are a welcome break from my other obsessions. If I am going to have my mind occupied, it should be by things of actual importance to me.

11:05am. Done chilling. It is time to start this thing.

If I am going to get sidetracked, it only makes sense that this be by the requirements of the path itself. I should not get jobs on the side - the path itself should sustain me. This is what I am going to test in 2021 most of all. I could not have much success aiming high, so what if I tried aiming low? If I can grasp the low, I can use it a stepping stone to higher places.

11:10am. I really want to give the middle finger to my general environment. I've been living here in the middle of nowhere for who knows how long, and I've had to endure being at the bottom for so long, but it was all for nothing in the end. Figuring out the meaning of life did not get me closer to anybody, instead those deep secrets just keep further estranging me from humanity. I don't know whose fault that is, but I know who is going to pay.

And what I must do until I get my power is embrace labor.

Well, I might regret it now, but it is not like it was not plan B all along. I have a nasty tendency to want to explore too much which just exarcebates my mental strain.

```fs
    // Typecase
    | ETypeLet of Range * Id * T * E
    | ETypePairTest of Range * bind: Id * pat1: Id * pat2: Id * on_succ: E * on_fail: E
    | ETypeFunTest of Range * bind: Id * pat1: Id * pat2: Id * on_succ: E * on_fail: E
    | ETypeRecordTest of Range * Map<string,Id> * bind: Id * on_succ: E * on_fail: E
    | ETypeApplyTest of Range * bind: Id * pat1: Id * pat2: Id * on_succ: E * on_fail: E
    | ETypeArrayTest of Range * bind: Id * pat: Id * on_succ: E * on_fail: E
    | ETypeEq of Range * T * bind: Id * on_succ: E * on_fail: E
```

What I need to do is get rid of this.

11:15am. Done. I've just chucked a large amount of code essentially overboard.

```fs
        | RawTypecase(r,a,b) ->
            let a = ty env a
            let id, env = fresh_ty_var env
            failwith "TODO"
```

This is where I need to start with once again.

```fs
| TMetaV of Id
```

Let me add this to the list of `Ty`s.

Also, I think I disabled opening the term scope in typecase right?

```fs
        let case_typecase d =
            let clauses d =
                let bar = bar (col d)
                let typecase = root_type {root_type_defaults with allow_metavars=true; allow_wildcard=true} >>= typecase_validate
                (optional bar >>. sepBy1 (typecase .>>. (skip_op "=>" >>. next)) bar) d
```

Yeah, that is the case. The truth is, as long as metavars aren't eagerly bound I could in fact allow it, but unless somebody demands this, I think I'll just leave it out. It is not important.

I am not sure if it was yesterday or the day before it, but I've had a great insight. I can do unification in the peval segment as long as I take care that metavars never leave the typecase.

Also, evaling the types never actually leads to nominals (which can have term segments) being unboxed, so I am safe there. All this combined means I can have an easy time of implementing what I want.

It is a pity. What I had previously would not have been bad if matching on regular functions was not necessary. The way I decided on doing it compiled style is because I decided I would not need it, but I was wrong. I absolutely do need to match on the results of regular type function application.

11:30am. Focus me, let me just get through the prepass parts. I need to introduce `ETypecase`.

```fs
| ETypecase of Range * T * (T * E) list
```

I guess I will set it to this.

11:35am.

```fs
        | RawTypecase(r,a,b) ->
            let a = ty env a
            let id, env = fresh_ty_var env
            failwith "TODO"
```

Let me start with this.

```fs
        | RawTypecase(r,a,b) ->
            let b =
                b |> List.map (fun (t,e) ->

                    )
            ETypecase(p r,ty env a,b)
```

Hmmm, this part is troublesome.

```fs
    and ty (env : Env) x =
        let f = ty env
        match x with
        | RawTMetaVar _ -> failwith "Compiler error: This should have been compiled away in typecase."
```

I actually want to reuse ty in order to convert the metavars.

I'll also have to return the env, so that I can eval the typecase cases.

```fs
    and ty' (env : Env) x =
        let mutable is_metavar : _ HashSet = null
        let mutable env = env
        let rec f = function
            | RawTMetaVar(r,a) ->
                if is_metavar = null then is_metavar <- HashSet()
                if is_metavar.Add(a) then
                    failwith ""
                else
                    failwith ""
```

I'll use a null here for efficiency.

11:50am.

```fs
    and ty' (env : Env) x =
        let mutable is_metavar : _ HashSet = null
        let mutable env = env
        let rec f = function
            | RawTMetaVar(r,name) ->
                if is_metavar = null then is_metavar <- HashSet()
                if is_metavar.Add(name) then
                    let id, env' = add_ty_metavar env name
                    env <- env'
                    TMetaV id
                else
                    env.ty.env.[name]
```

This is pretty good.

12pm. Now I am just getting lost in thought. Let me complete that case.

```fs
        | RawTypecase(r,a,b) ->
            let b = b |> List.map (fun (t,e) ->
                let env, t = ty' env t
                t, term env e
                )
            ETypecase(p r,ty env a,b)
```

This should be good enough.

Yeah, this is good enough for the morning session. Later I will fill out all the prepass cases. Then I'll move to the peval. It should not be too hard. I can do this thing. If nothing goes wrong, I should manage it today. Let me have breakfast here."