"11:05am. Yesterday I had to stop 5 chapter till the end of vol 5 as I was too tired. Let me read them now.
12:05pm. That was too good. I love how every volume end is such a slaughterfest. I am going to give Reverend Insanity a pause for a month. 1022 chapters is enough.

12:25pm. Let me have breakfast here.

12:55pm. Let me do the chores here. After that I'll get started on the docs.

1:40pm. Done with chores. Now I can finally get started.

Let me turn off the router.

1:45pm. I need to gather my thoughts for a bit. The next part of the docs is serialization. I need two different kinds - one for regular pickling and one that is like the scheme I used in the old Spiral for passing data to the NNs.

This will take language changes so I am feeling pressured. I won't be able to do all of this today. Instead I will have to give it more time.

I need to steel my mind. Forget the future. Forget the chips. Once I can do serialization in all its forms, it is then that Spiral will be truly complete.

1:50pm. I keep thinking back to GADTs and am feeling confused as to why I decided to leave them out. Only by looking back in the journal have I been reminded about the impossible cases. Yeah, now I remember it. GADTs would require dependent pattern matching. I completely forgot about this. Great, now I can put it out of mind again.

1:55pm. I am going to have to redesign the language a little in order to make it more capable at serialization.

2pm. I really do not feel like starting work right away at all. I just want to sharpen my intent today.

After I do serialization, everything that I wanted to cover in the docs will have been covered. Past this point there should be no more extensions to the partial evaluator. And the language has been put through its paces.

I am making strong progress.

v2 is indeed harder than v0.09. If I had to recreate all of this from scratch it would take me at least 3 months. But even without experience, the design is good enough that it can be considered a man year of work. This meets one of my goals. If Spiral was 10 man years, it would not be tractable.

2:05pm. Let me take a short break here.

2:20pm. I am back.

2:25pm. Why am I having trouble starting here?

Because docs are almost over. After the next step is made, it will almost be time to start the sponsor search. I'll deal with serialization, but putting in the comments is an UI change instead of a language change.

2:30pm. Sigh. I started the docs 1.5 weeks ago, and said that it would be a long journey that would take the whole of January, but now I can already see the end in sight.

Let me go to bed for a while here. Then I will gather my resolve and start work on serialization. Once I start, it will become my singular obsession until it is done to satisfaction. I won't cut any corners here.

I've been really been indulging myself lately and it has slackened my spirit. I need to take some time off from the screen in order to remember the desire that drives me.

2:40pm. I feel a restlessness that fatigues me. The works becomes routine. At first it keeps you going, but then it becomes a harpy screaming at you that unless you work, you are worthless.

Every once in a while, it is good to take some time off in order to remember why you are working in the first place.

All this time I've been working on the docs, I've been working on the language, and I've been thinking what my principles should be when it comes to future social interaction. But today I should take some time to internalize the feeling of accomplishment."