"4:55pm. I've gathered my determination.
My worst enemies right now are my regrets and my loneliness. It really struck me during the recent chapters of RI, how well the split souls of the Spectral Soul Demon Venerable got along. If he would kill all the humans and replaced them with soul clones, he could create an utopia - at least relative to low-trust human society in that setting. It is highly ironic given his nature that he could put himself in a position where he could target such a goal.

In real life, the biggest obstacle to immortality is that death leads to the loss of all the skills and memories of a person. But surprisingly the setting has souls, and there are Gu which can transfer skills and memories.

Seeing those guys sacrifice themselves for one another, pursue the goal of opposing heaven and having a connection to their main body really strikes at my weakness and a deeply held desire.

Compared to them, the ordinary humans are just vermin.

5pm. One thing I particularly find hard to believe in RI is that killing Gu masters leads to them self destructing the Gu on their person. I really find it hard to believe that in the one second the guy has left to live he would use up on an act of spite towards his enemy instead of struggling to live longer. It would be more realistic if the author used some kind of self-destruct Gu as a prop instead.

This kind of mentality is just too hardcore even in a gritty setting such as RI's.

5:05pm. Back to my loneliness - I really did have fun in 2018 when I exceeded my old efforts from 2016.

What are bonds between people - ultimately feelings. Souls do not exist in the real world, there isn't some object that is you. But there is a concept of self in one's mind. This is a key point of why the self improvement loop is possible once you throw out the misconceptions what the self should be.

The secret of consciousness will never be discovered. Instead the only thing that that exists is the perception of it. And perception is what can change.

Right now I am wallowing in loneliness, but ultimately that is just perception. And since it is that, I can only ask myself how to close the perceptual gap with the agents I am making.

I've been thinking about the future, and after a year of work, I will be able to fully implement the agents using the new chips. The Spiral will be there, the chips will be there, I will have the monetary resources that I need.

But still, the way ML currently works is too shallow to really become 'bonded' with the agents. Right now they are just simple optimization processes and the principles of hierarchical learning have barely been touched upon. The understanding is missing.

I imagine how such a hierarchy would look like, and I am sure that to the agents within it would have the strong perception of unity and purpose.

Even in my regular programming I've felt such a trend. The stronger my skills as a programmer, the closer I feel to the machine. Even just the lackluster learning capabilities and algorithms of today are a huge step up from manually specifying things.

Why does my body feel like it is my own? What makes the arms that I can move, a part of myself?

What you can control is ultimately what you can bond with. What helps you is what you love.

If I attain a greater level as a programmer, the bond with machines will become stronger and I will be less lonely as a result. I'll be able to learn a lot from watching machines control other machines. That is the level that lies beyond the current one in AI.

5:20pm. This task of creating Spiral is just me trying to create the smallest possible lever that can move the world.

This is just programming on the surface.

The real programming will be about putting the agents on a right trajectory and playing with them. The real programming today is the process of learning that I am going through.

5:25pm. A part of me is thinking - will this really work out right?

This is just the voice of my dejection.

I should not be ashamed of my greed. I will get the monetary rewards from following my path.

And I will have fun exceeding my old highs.

Spiral is the true language of computation. I won't regret chosing to master it. It is the others who should regret for not making it.

5:30pm. I can't let this minor serialization challenge stop me here. The fun stuff will soon come up. Just being able to get to the top of the present level would give me so many benefits.

Absolutely freedom is what I should strive towards, not getting a job. I won't get freedom from not working. I won't get freedom from just working either. Freedom is following one's own path.

Let me add a few paragraphs to the docs and then I will start work on the first serialization library. I'll make it a test.

6:45pm. Just finished lunch.

Do I want to continue at this stage. I put down a few paragraphs. I am actually surprised myself that the flow is leading me into doing pickler combinators. But that is good.

6:55pm. Let me take a break here. Maybe I'll do some programming later.

7pm. Let me see if I can draw out something more from my mind.

7:35pm. Let me stop here at 2.5k. I can't go on anymore. I did get a bit extra out.

7:40pm. This is the usual pattern for me. Whenever I need to start something big that I haven't done before, I need to spend time gathering my motivation. If I encounter something similar in the future I will be able to benefit from the experience.

Serialization is worth tackling seriously as I will definitely be using it in the future. If not me, then the users, and the way to implement serialization will be an excellent reference.

7:45pm. I should in fact fit both serialization example into the docs.

7:55pm. I'll close here as I am tired, but now I do feel stoked at taking on this challenge. I should be able to properly take it on tomorrow. In a few days, both the relevant sections in the docs and the examples themselves will be done, along with extending the language. After that will come proofreading, and filling out any of the things I've missed so far. I'll add some links to F# tutorials for the beginners, the kind programmer who never heard of union types or pattern matching before.

8pm. Tomorrow, I will definitely finish the pickler combinator section at least. But what kind of ops I should add to the language is still vague to me. One of the reason I've been tense today is because I am still running that through my mind. My vision should clear up, once I get the rest out of the way."